Add `aph/open-line` and `aph/newline`
These functions tack the functionality of `join-line` into the unused
negative-argument space of `open-line` and `newline`, respectively
(one for the usual behavior, and the other for the "with argument"
behavior).

The function `aph/open-line` is bound to `C-o` (instead of
`open-line`), but `aph/newline` is not yet bound to any key.