Make `C-h a` a prefix key for several apropos bindings
The `apropos` command, which was formerly at `C-h a`, is now at
`C-h a a`.  Other commands on this prefix are:
- `apropos-command`, at `C-h a c`
- `aph/apropos-function` (see below), at `C-h a f`
- `apropos-variable`, at `C-h a v`
- `apropos-value`, at `C-h a <space>`
- `apropos-documentation`, at `C-h a d`
- `apropos-library`, at `C-h a l`
- `info-apropos`, at `C-h a i`
- `tags-apropos`, at `C-h a t` (pending future tags support)

The new function `aph/apropos-function` is just like `apropos-command`
with an inverted prefix argument (that is, it shows all functions by
default and just commands with a prefix).

We also establish alternate bindings for the Cider apropos functions,
at `C-h A a` for `cider-apropos` and `C-h A d` for
`cider-apropos-documentation`.  These are only active in Clojure mode,
and the intention is for all apropos commands specific to languages
other than Emacs lisp to reside under the `C-h A` prefix key on their
individual mode maps.

Finally, we let `apropos-do-all` return to its default nil binding, as
having it set reduces the amount of flexibility available with these
apropos bindings.