Begin migrating to a personal custom theme
This commit improves visibility of the region.  To accomplish this, it
was necessary to override choices made by the themes I use ('zenburn
and 'hc-zenburn), and the cleanest way to do that seems to be making
my own theme extending those two.  The file aph-theme.el now contains
that theme; the theme-switching functionality it used to contain has
been disabled and moved to init-draft.el until it can be reworked to
fit the new way of doing things.