New command `aph/info-mode-or-clone-buffer` bound to `C-h i`
This command extends `info` (the default binding for `C-h i`).
Outside info mode, or if a prefix argument is supplied, both commands
behave identically.  Inside info mode, the new command will clone the
buffer; the old command did nothing in this circumstance.

The advantage here is that the key `C-h i` can absorb the
functionality of the info mode key `M-n`, which is currently shadowed
by `smartscan-symbol-go-forward`.