Add confirmation prompt to `aph/delete-frame-or-exit` (`C-x C-c`)
Because I have Windows configured to only show me the frames on a
particular desktop, but Emacs considers frames on all desktops to be
visible, it was too easy to hit `C-x C-c` intending only to close the
current frame and instead quit Emacs entirely.  Adding a confirmation
prompt for that case fixes the problem.