Enable use of `smart-tab` inside Org-mode
Advise `org-cycle` to use `smart-tab` as its fallback action instead
of just indenting.  This allows use of `smart-tab` without interfering
with any of the tab key functionality specific to Org-mode.

Features still to be implemented:
- Use Org-mode specific completion types (e.g., TODO keywords).
- Allow completion inside table cells.

Note that `smart-tab-mode` is still not enabled in Org-mode; for
clarity's sake (and to allow an easy way to disable `smart-tab` inside
Org-mode), we may want to rework our implementation so that it's only
in effect with `smart-tab-mode`.