Set up `company-mode`
The rough intention is to use `company-mode` in programming modes and
continue to use `smart-tab` with `hippie-expand` in other modes.  At
present, `company-mode` is enabled only in Lisp modes.

The configuration for `company-mode` has been changed from the default
in the following ways:
- Idle completion is disabled entirely.
- The tab key has been rebound to the `smart-tab`-like command
  `company-indent-or-complete-common`.
- When the completion overlay is present, tab is bound to
  `company-complete-common-or-cycle`, which is as close as
  `company-mode` gets to `hippie-expand`.

On an unrelated note, several instances of `eval-after-load` have been
replaced with `with-eval-after-load` (which is simpler to use, as it
does not require quoting or `progn` forms).  The intention is to
gradually replace all occurrences over time.