New command: `aph/Info-final-menu-item`
This `Info-mode` command follows the last menu item in the current
node, duplicating the functionality of the `0` key in the standalone
`info` application, which was inexplicably left out of at least the
current version of Emacs.  It is still bound to `0`.

Since this is the second command for `Info-mode` I've produced (the
first being `aph/info-mode-or-clone-buffer`), it seemed like a
reasonable time to split those functions off into their own file,
aph-info.el.