Begin drafting code to adjust Elfeed's browser usage
We want Elfeed to use `eww` as much as possible.  The new draft code
works toward that end, although I still haven't worked out the desired
behavior exactly.

As a first step, the draft code includes a function `aph/browse-url`,
which will be a replacement for `browse-url` at `C-c C-o` once
Org-mode can be convinced to use `eww` by default as well.  This
function opens the specified url (defaulting to one at point, if there
is one), usually in `eww` but in the default external browser with a
prefix argument.