Add commands `aph/elfeed-show-next` and `aph/elfeed-show-prev`
These commands extend `elfeed-show-next` and `elfeed-show-prev` with
the functionality of `aph/elfeed-search-show-entry`.  They replace
those commands, at `n` and `p` respectively, in Elfeed entry buffers.

Additionally, these commands are bound to the same keys in `eww`, and
the commands `eww-next-url` and `eww-previous-url` (which previously
occupied the `n` and `p` keys) have been moved to the
previously-unused `]` and `[` keys.