Add single-use support to `aph/with-advice`
Prefixing an advice clause in `aph/with-advice` with the keyword :once
now instructs the macro to install that advice using `aph/advice-once`
instead of `aph/advice-add`, so that it is executed at most once.

One minor issue: if the advice is not executed, the cleanup handler
installed by `aph/advice-once` remains until it is executed.  It
should not interfere with actual functionality except in particularly
exotic circumstances, but it will clutter up the `C-h f` view.