Add and configure Helm
We are now using Helm for completion, replacing Ido.  This commit
makes the replacement, configures some Helm settings, and updates many
keybindings.  Most of the changed keybindings are Helm-related, but a
few are not.  In particular, we begin using the `C-z` prefix to launch
processes which might be considered "subapps" of Emacs (e.g., `eww`
and `elfeed`).