Split up commands on `C-x r` prefix
The `C-x r` prefix is by default pretty disorganized.  It contains the
register commands, the rectangle commands, and the bookmark commands.
In order to make the keys more memorable, and to provide more space
for bindings, we are distributing the commands among three separate
prefixes: `C-x r` for register commands, `C-c r` for rectangle
commands, and `C-c b` for bookmark commands.