Install `helm-descbinds` package
We are now using the `helm-descbinds` package.  This enables use of
Helm to view keybindings on a prefix, via the usual `<prefix> C-h`
idiom.

Since viewing the `C-x` prefix in this way is now manageable, we also
move the Helm command prefix (formerly at `C-x C-h`) to `C-x x`, which
is currently unbound.