Make `shr` not underline image links
Underlining image links causes problems with `shr-zoom-image`, so we
tell `shr` not to underline these links (and only these links).
Ideally, we would add some other way to differentiate these links from
ordinary images, but for now this will do.

This affects `eww` and `elfeed`.  We also remove the 'link tag from
the Questionable Content feed, as we only added this tag as a
partial workaround for this issue.