Implement favorite filters in Elfeed
Add the new function `aph/elfeed-search-next-favorite-filter` (bound
to `'` in the Elfeed search buffer).  This function cycles through the
filters specified in `aph/elfeed-favorite-filters`.

To support this function, we also introduce a new utility function in
aph-lib, `aph/successor-in-list`.  This function takes a list and an
element and returns the successor to that element, if it appears in
the list.  It can also take an optional argument that causes it to act
cyclically, i.e., to return the first element of the list when given
the last element.