Tweaks toward conditional package loading
Since MPC is not configured to load packages itself, we need a way to
allow that machine to opt out of `use-package`'s' `:ensure`
directives.  This is complicated by the fact that `:ensure` does its
work as part of macro expansion (!), so it's difficult (maybe
impossible?) to make those directives machine-aware.  For the moment,
we'll just avoid using `:ensure`, but it may be worth leaving in some
of the code we were working with, particularly the code that
bootstraps `use-package`.