Move smart agenda code to aph-org-agenda.el
Move `aph/org-agenda-display-smart-agenda` and associated variables
from init-org-agenda.el to aph-org-agenda.el.  Also, rework this code
so it does not depend directly on the custom agenda views defined in
init-org-agenda.el.  Instead, let the agenda views to be used be
customized in a variable, and leave that customization in
init-org-agenda.el.