Rewrite code updating Org-mode markup and move to aph-org.el
Replace the function `aph/org-element-text-markup-successor-advice` in
init-org.el with the more general `aph/org-emphasis-alist-update` in
aph-org.el.  This allows us to separate the configuration code in
init-org.el from the implementation details in aph-org.el, in
concordance with the ongoing `use-package` refactoring.

The new function `aph/org-emphasis-alist-update` just takes a
single-character string identifying one of the markup delimiters from
`org-emphasis-alist` and does everything necessary to replace it with
another delimiter.  This includes advising
`org-element-text-markup-successor`.