Begin drafting mode tag feature
This feature is intended to simplify the ongoing `use-package`
refactoring by allowing me to attach hook variables to arbitrary
collections of modes.

So far, `aph/mode-tag-create` should be fully functional but requires
more tests to be written.  The other functions exist only as stubs.
Naturally, the module (aph-mode-tag.el) is not loaded on
initialization.