Continue drafting mode tag functionality
The principal advance in this commit is the development of the macro
`aph/with-test-mode-tag`, which should make writing tests for this
module much more straightforward.

In addition, the macro `aph/mode-tag-create` has been completed, and
the logic creating the hook variable name from the tag name has been
spun off into the function `aph/mode-tag-hook-var`.