Draft new function `aph/symprop-delq`
This relatively simple function updates a list stored as a symbol
property, deleting a specified element using `delq`.  This process is
rather verbose, and implementing it as a distinct function allows us
to make our implementation of `aph/mode-tag-remove` much cleaner.

To house this function, we have created a new module, aph-symprop.el.