Draft function `aph/symbol-concat`
This is a very simple function, but the testing module for
`aph/mode-tag` uses this operation repeatedly, and it seems worthwhile
to break it out as its own function.

The new function resides in the module `aph-symbol.el`, which was
renamed from `aph-symprop.el` upon receiving its new wider mandate.
This module will house all kinds of functions for manipulating
symbols, not just those involving symbol properties.