Make tagged modes run their tags' hooks
Tagged modes now run all hooks for the tags they are associated with.
This is accomplished via the new function `aph/mode-tag-run-hooks`,
which is automatically added to the mode hooks of tagged modes by
`aph/mode-tag-add` (and removed by `aph/mode-tag-remove` when a mode
loses its last tag).

With this commit, mode tags seem to be fully functional, and the
`use-package` refactoring can resume.