Fix mode tag inheritance issue
Change how tag hooks are run.  The previous implementation had the
problem that tag hooks for inherited tags weren't run, and the obvious
fix had problems with doubly-inherited tags (which shouldn't happen
all that often, but is certainly possible).

The new implementation runs `aph/mode-tag-run-hooks` once for every
mode, whether it has any mode tags or not.  Since this function does
nothing when a mode is untagged, this is safe.  To accomplish this,
the function is added globally to `change-major-mode-after-body-hook`
rather than the individual mode hooks.