Move most `M-=` prefix bindings into init-core.el
Most of the commands on the `M-=` prefix are from the `simple`
package, including the command that resides at `M-=` by default,
`count-words`.  This commit moves the keybinding code for these
commands from init-keys.el to init-core.el.  It also cleans up the
`use-package` declarations for the `help+` family of packages.

Moving the remaining `M-=` bindings may be tricky, as these will go in
different `use-package` declarations.  The naive solution (just moving
the individual bindings) errors out, because the `M-=` prefix may
still be in use by `count-words`.  The next-most-naive solution is to
unbind `M-=` in each `use-package` declaration where the prefix will
be used, but this doesn't work because subsequent unbindings
annihilate any existing bindings on the `M-=` prefix from other
declarations.  Further investigation is warranted, and the final
solution will most likely involve the `:prefix` or `:prefix-map`
keywords for `bind-keys`.