Move binding for `M-= b` into init-core.el
This binding establishes an alternate keybinding for the command
`describe-buffer` from the package `help-fns+`.  Because the other
commands on the `M-=` prefix are in a different `use-package`
declaration (for the `simple` package) and there is a command bound by
default to the prefix key `M-=`, coordinating the bindings is somewhat
tricky, especially since we want to find a general solution that does
not depend in any way on the ordering of the individual declarations.

The cleanest apparent solution is to explicitly define the keymap as a
variable (`aph/buffer-info-map`) before the `use-package`
declarations.  A function alias at `aph/buffer-info-prefix` is also
established there.  Individual bindings are then made to that keymap,
and the key `M-=` is rebound to the keymap itself in the `use-package`
declaration for `simple`.