Move `s-<apps>` bindings into init-core.el
Four bindings were moved from init-keys.el to init-core.el
- `s-<apps> d` for `delete-duplicate-lines`
- `s-<apps> s` for `sort-lines`
- `s-<apps> k` for `flush-lines`
- `s-<apps> M-k` for `keep-lines`

These bindings belong to two packages, `sort` and `replace`, and the
latter is not a true package, as it has no `provide` statement.
Moreover, the `s-<apps>` prefix is one I intend to use quite widely.
Hence we follow the same procedure used with the `M-=` prefix and
explicitly define the keymap as `aph/region-manip-map` and
`aph/region-manip-prefix` (variable and function, respectively).