Move binding for `C-z` prefix into init-core.el
Replace the code in init-keys.el unbinding the `C-z` key (and
therefore making it suitable for use as a prefix key) with code in
init-core.el that assigns this key the new prefix map
`aph/launch-prefix`.  This is consistent with the way we've handled
similar prefixes in recent commits.

The new code is in the `use-package` declaration for the `frame`
module because that's where the old binding for `C-z`
(`suspend-frame`) is defined.  This seems like a reasonable enough
place to put it even though we never reference anything actually
defined in that module.  Since this is a built-in module that will be
loaded on startup anyway, there is no problem with loading it
unnecessarily.