Move `ielm` keybindings into `use-package` declaration
Move the following two keybindings (both local to `ielm-mode`) from
init-keys.el into the `use-package` declaration for `ielm` in
init-core.el:
- `C-c M-w` for `aph/ielm-copy-last-output`
- `C-c C-t` for `aph/eval-expression-toggle-clean-output`

By rights the keybinding for `C-c C-t` should maybe be in the
declaration for `aph-simple`, but that would require the use of
`with-eval-after-load` to defer loading until the keymap `ielm-map` is
defined.  Instead we add an autoload for it; this is probably
unnecessary since `aph-simple` should be loaded on initialization.  At
the same time we remove the unnecessary autoload on
`aph/ielm-copy-last-output`.

We also reinstate the advice enabling
`aph/eval-expression-toggle-clean-output`, which was apparently
deleted by mistake earlier in this refactoring.