Configure `mouse-avoidance-mode`
Configure mouse avoidance mode to move the mouse out of the frame
entirely, out of the bottom-right corner.  Since I'm using Emacs
entirely on Windows, this should be safe, but this may cause issues if
I ever try to use it on Linux systems that have window focus follow
the cursor.

The other concern is that pushing the mouse cursor out the
bottom-right (as we are) will trigger the "hide all windows" hover
area on the right side of the taskbar, but testing shows that this
does not appear to be a problem; either the coordinates I have chosen
avoid that area or else Windows is able to detect that the mouse is
being moved programattically and avoids triggering it).