Implement workaround for `org-agenda-quit` bug
The command `org-agenda-quit` appears to be bugged; executing the
command does nothing.  Tracing execution suggests that `bury-buffer`
may have picked up a bug at some point.  (For reference, the current
versions are Emacs 24.5.1 and Org 8.3.2.)

As a workaround, the command `quit-window` seems to work fine, and
none of the additional behavior of `org-agenda-quit` seems to be
relevant for me, so I'm just going to use that for the time being.