Move `aph/number-lines` and related commands to `aph-number-lines`
Move the commands `aph/number-lines`, `aph/number-lines-alpha`,
`aph/number-lines-open`, and `aph/number-lines-open-multiple` from
`init-draft` to the new module `aph-number-lines`.  Move associated
tests to the testing module `aph-number-lines-test`, and move
keybindings to a `use-package` declaration in `init-core`.

In the process, we noticed and fixed a small bug in `aph/number-lines`
that caused the formatting to be off when using the default format
string and there are exactly 100 lines in the buffer.