Move `rectangle-number-lines` from `C-c r N` to `C-c r n`
This command was originally at `C-x r N` before we split up the
bindings on the `C-x r` prefix.  It couldn't use `C-x r n`, because
that was (and still is) bound to `number-to-register`.  After the
split, there's no reason to keep the capital letter, and this will
also be more consistent with `aph/number-lines`, on `S-<apps> n`.