Fix bug in `aph/silence` regarding nil messages
The function `aph/silence-advice` was generating errors by passing nil
as an argument to `format`.  This is because `message` calls with
first argument nil or "" are treated specially and are used to clear
the current message from the minibuffer.  This has been fixed;
`aph/silence-advice` will now ignore messages of this type and will
never attempt to silence them.  More importantly, it won't try to
`format` them.