Make `helm-projectile-grep` suspend updates initially
The command `aph/helm-projectile-grep` is intended to be a drop-in
replacement for `helm-projectile-grep`, currently at `M-s M-g`.  This
command acts in the same way, except helm updates will be suspended
until the user presses enter (bound to the new command
`aph/helm-resume-update-or-exit-minibuffer`) once, at which point
updates will resume and the helm buffer will not close.

Note that this latter behavior (resume updates rather than exit) is
global to all helm buffers.