Implemented `aph/mpc-calc-bar`
This function sets up a short window of the sort which I've been using
at work, with `calc` on the left and `ielm` on the right.  It is bound
to `C-z C-=` on MPC.

As part of developing this function, I discovered several apparent
inaccuracies in the way frame geometry was being reported, probably
due to idiosyncracies with the Windows OS.  For the present
application, I dealt with this by implementing my own parallel
functions in the module `aph-frame`.  I'm not sure whether this was
the correct approach or not.

Finally, note that this commit introduces the macro
`aph/save-frame-excursion` in `aph-subr`.  This does exactly what it
sounds like: execute its body, then restore the original frame
configuration.