Fix Cygwin compatibility problems with Haskell mode
Because Stack is outside of Cygwin but Emacs is inside, Haskell mode
needs to be doing some additional path translation when interacting
with GHCI.  This commit implements that translation, in the form of
`aph/haskell-process-load-file-cygwin` (a replacement for
`haskell-process-load`).  Also the `^M` characters in the REPL process
log were screwing things up (not to mention being ugly), so we make
sure we're using DOS coding for this.