New function `aph/call-logging-hooks`
This function calls the specified command and reports what hooks were
run during its execution.  It was originally written for a
StackExchange answer, but has sufficient utility to be incorporated
into my init files.

Currently this function resides in the `aph-help` module, because it
is focused on providing introspection information about Emacs.  But it
doesn't seem to belong there, so I may move it at some future point.