New command `aph/helm-forms-list-records`
This command is designed to be invoked from within a `forms-mode`
database and allows the user to jump to any record in that database.

The name displayed for each record in the helm buffer is the value of
field number 1 for that record.  At present there is no easy way to
change this; future commits will improve the functionality.