Liberate `C-[` key from ESC
Make the `C-[` available for binding by severing its legacy
relationship with the escape key.  We are careful to make sure the fix
works with `emacsclient` by using `after-make-frame-functions`, and we
maintain the association between the escape key and the meta key
(since doing otherwise would risk making emacs unusable).