New command `aph/move-beginning-of-line`
This command combines the functionality of `move-beginning-of-line`
(`C-a`) and `back-to-indentation` (`M-m`), exploiting the fact that
`move-beginning-of-line` does nothing when already at BOL and there is
no prefix argument.

This is intended to replace `move-beginning-of-line` at `C-a`, but
that keybinding has not been done yet, as this is a good test case for
`aph-keys-mode`, currently in draft.