New function `aph/equal`
This function extends the equality predicate `equal` by allowing the
comparison of any number of arguments (like in Clojure).

It is currently housed in the `aph-dash` module because it is
implemented using `dash` functions, but this is more arbitrary than
usual, since the base `equal` predicate is a built-in function.