New testing macro `aph/ert-macro-executes-body`
This macro tests whether a supplied macro executes a body and takes on
its return form (in the manner of `progn`).  It will allow us to
refactor this kind of simple functionality test for macros of this
sort and avoid code duplication.

At present, the only two tests of this type are in the
`aph-mode-tag-test` module: `aph/mode-tag-test-with-mode-tag--body`
and `aph/mode-tag-test-with-mode--body`.  These have been refactored
to use the new macro.