New function `aph-keys-augment`
This function is intended to provide support for mode-local bindings
within `aph-keys-mode`.  At present, all it does is create an
"augmented keymap" for a major or minor mode (really any symbol, but
only mode-based support is planned at the moment).  Keybindings can be
added to this keymap, but the code required to activate the keymap at
the proper time has not yet been written.

Note that `bind-key` should not yet be used with an `aph-keys-augment`
call (or any kind of sexp), since `describe-personal-keybindings`
can't handle anything except a plain variable.  I intend to add
support for this kind of call shortly.