Allow `aph/update-alist` to use `assq`
Add an optional argument to `aph/update-alist` for a function to use
in place of `assoc` to find the desired association.  This allows use
of `assq`, for instance.  Also add a macro `aph/set-assq`
corresponding to `aph/set-assoc`.

In addition, allow `aph/set-assoc` (and the new `aph/set-assq`) to
operate on any generalized variable.