Make bindings automatic in `aph/ert-with-test-mode`
Previously `aph/ert-with-test-mode` accepted a list of binding names
of the form `(mode hook keymap)`, with both `hook` and `keymap` being
optional or semi-optional (since it was impossible to supply `keymap`
without also supplying `hook`).

This (rather clunky) system has been replaced by a single binding
name, from which the others are automatically derived.  This is
equivalent under the old system to passing the list `(mode mode-hook
mode-map)`, paralleling the actual names for the underlying variables.