Rework handling of major modes in `aph-keys-mode`
The previous method `aph-keys-mode` used to handle an augmented major
mode involved making the variable `minor-mode-map-alist` buffer-local,
which turned out to be very fragile.  In particular, it prevented
`edebug` from setting up its keybindings correctly.

The new approach stores the augmented local map in a second variable
in `emulation-mode-map-alists`, called `aph-keys-local-map-alist`.
This variable is separate from `aph-keys-augment-map-alist` so it can
be buffer-local, and we set up `emulation-mode-map-alists` so that it
follows `aph-keys-augment-map-alist`, thus preserving the usual
precedence order of keymaps (minor -> major -> global) when looking
only at augmented keymaps and `aph-keys-mode-map`.  (Of course,
unintentionally overriding other modes' mappings may still be a
concern).