Add support for derived mode parentage to `aph-keys-mode`
Binding a key in the augmented map for a major mode will now use that
binding in `aph-keys-mode` with all major modes descended from that
mode, unless shadowed by a descendant mode (or a minor mode, of
course).

This attempts to mimic the keymap inheritance for major modes, but
since it is following mode inheritance rather than keymap inheritance,
augmented keymaps may inherit differently than the base keymaps for
their underlying modes.  It isn't clear yet whether this issue needs
to be addressed.