Fix startup bug related to function `get-mode-local-parent`
The function `get-mode-local-parent` that I was using to look up the
parent of a derived major mode was causing a startup error, as it is
not autoloaded.  After further investigation, it appears that this
belongs to a library that attempts to add mode-local variables, and as
such is probably not the right tool for my purpose.  I have replaced
it with a direct lookup of the symbol property `derived-mode-parent`.