Add stopgap fix for `elfeed` issue involving `aph-keys-mode`
Since `elfeed-search-mode` does not follow the convention of running
its mode hooks with `run-mode-hooks`, it does not run
`after-change-major-mode-hook` and thus `aph-keys-mode` does not have
a change to update the augmented major mode keymap.  For the moment,
adding `aph-keys--update-major-mode` to `elfeed-search-mode-hook` is
an easy fix.  In the long run, we may be able to fix the problem using
advice if it doesn't get fixed by the `elfeed` developer directly.