Update `aph-keys-minor-map-alist` when augmenting new keymaps
When a keymap is freshly augmented, it is added to
`aph-keys-augment-map-alist`, but the variable used to detect whether
an augmented minor mode map should be active is
`aph-keys-minor-map-alist`; if `aph-keys-mode` is active, this won't
be refreshed from `aph-keys-augment-map-alist` until `aph-keys-mode`
is toggled off and back on.  With this commit,
`aph-keys-augment--register` now updates this variable directly
immediately after adding the keymap to `aph-keys-augment-map-alist`.