Remove prefix maps and clean up binding code
The prefix maps we were using previously to collect global bindings
from different packages (e.g. `aph/region-manip-map`) are no longer
necessary, as these can all be moved into `aph-keys-mode-map`.

The idiom of using `bind-keys` in the `:config` section of a package
declaration (so it is run after the mode map is defined) is also
unnecessary now, as there is no problem with binding keys in an
augmented keymap before its base keymap is defined.  We can move all
of these statements into a `:bind` form.