Bind `C-m` to `copy-to-register`; sort out <return> issues
Bind `C-m` to `copy-to-register`.  This will be replaced shortly with
a more sophisticated DWIM-style command, and aligns broadly with my
desire to make the register commands easier to use.

However, the association between C-m and <return> in Emacs makes this
process less straightforward than I would have hoped.  In particular,
many (all?) modes define the binding they want for <return> at RET and
rely on the underlying translation.  Thus, binding a key to `C-m` in
any active keymap will override the command at `<return>`, unless we
bind something to `<return>` as well.  The general solution is the new
function `aph-keys-default-return-command`, which just executes
whatever command would be bound to `C-m` if `aph-keys-mode` weren't
interfering.