Fix `calc` bug related to `aph-keys-default-return-command`
Entering numbers in `calc` has exhibited some unexpected behavior
since the introduction of `aph-keys-default-return-command`.
Specifically, pressing `<return>` in the minibuffer to finish entering
the number caused the number to appear twice in the `calc` buffer.

After some investigation, this appears to be due to a hard-coded check
for `C-m` in the `calc` minibuffer code.  The solution was
straightforward: `let`-bind `last-command-event` to `C-m` around the
call.

A first attempt to fix this involved passing `C-m` as an optional
argument to `call-interactively`; this didn't fix the problem, but
didn't seem to cause any harm either, so I'm leaving it in on the
theory that it might quietly fix a different compatibility problem.