Sort out problems with `M-p`/`M-n` in `comint` buffers
The first problem is that the bindings we have established for `M-p`
and `M-n` (`backward-paragraph` and `forward-paragraph`) are shadowing
the default `comint` buffers; the solution for this is to make those
bindings local to `prog-mode` and `text-mode`.

Once those bindings are out of the way, it turns out that
`smartscan-mode` is also shadowing these same bindings, using its
default bindings.  The solution for this is to not enable
`smartscan-mode` globally, but instead turn it on in the `prog-mode`
and `text-mode` hooks.