Overhaul `smartparens` keybindings
This is a major overhaul of all of the `smartparens` keybindings.

As part of the overhaul, I encountered problems binding
`sp-forward-barf-sexp` to `C-M-[`, related to the `C-[`-ESC
connection.  The issue is that `C-[` is equivalent to `M-ESC` and thus
to `ESC ESC` and is therefore a prefix of the `ESC ESC ESC` "get me
out of here" binding.  The solution is to move that binding in
`aph-keys-mode-map` to the key we actually want to use for that
purpose, `<escape> <escape> <escape>`.