Add `bind-keys` support for overriding keymaps
Add the :override keyword argument to `bind-keys`; this argument works
exactly like :augment, but uses the overriding augmented keymap
instead of the usual, non-overriding one.

In addition, rework our handler for the :augment and :override keymaps
so both keywords accept either singletons or lists, and both are
compatible with the :map keyword.  (Previously, the use of :augment
would shadow :map).  Thus, defining a binding for multiple (augmented
or unaugmented) maps is now possible in a single `bind-keys` form.