Begin extending `outline` mode
Preparatory to writing a function that descends in the `outline-mode`
hierarchy (an inverse to `outline-up-heading`), I am writing some
useful subroutines.

The first is `aph/outline-before-first-heading-p`, a generic analogue
to `org-before-first-heading-p`.

Note that, at present, I am not testing these subroutines against
invisible headings, as I feel that the work-benefit ratio for that is
too low.  If I run into problems, I'll add regression tests as I fix
them.