Split `aph/outline-before-first-heading-p`
Rename the function `aph/outline-before-first-heading-p` to be
`aph/outline-before-first-heading`, and add a new function
`aph/outline-before-first-heading-p` that has the same return value
but does not change the match data; this is analogous to the
difference between `looking-at` and `looking-at-p`.

The reason for this is my newfound understanding that the
function-variable `outline-level` needs the match data to be set
before it can be used, and that most of the `outline-mode` functions
set this data.  I would prefer to isolate this state as much as
possible in my own code, so I need to make the distinction as
carefully as possible.