Fix problematic interaction between `aph-keys-mode` and `helm`
Fix problem wherein `aph-keys-mode` was shadowing `helm-map` bindings
(most noticeably, `s-<apps>`).

To accomplish this, we add a new keymap `aph-keys-mode-helm-map` which
is used in preference to (and inherits from)
`aph-keys-mode-minibuffer-map` when `helm` is active.  This keymap is
used in place of the augmented map for the current major mode and has
precedence over `aph-keys-mode-map`, so making the relevant bindings
in this keymap fixes the precedence problem.