Work around lack of `use-package` support for complex :bind
The macro `bind-keys` supports multiple arguments to the `:map`
keyword (and my extension uses this to add support for multiple
arguments to `:augment` and `:override`), but `use-package` does not
support this usage properly, and this was causing some keybindings not
to be made (most notably `C-i` and `C-M-i` for `align` and
`align-regexp`, respectively, in `prog-mode`).

For the time being, I have replaced each of the offending `:bind`
statements with a sequence of single-map `:bind`s.