Update command `aph/delete-frame-or-exit`
Make the command `aph/delete-frame-or-exit` (`C-x C-c`) better reflect
my usage patterns with respect to `emacsclient`.

Because the previous implementation used `save-buffers-kill-terminal`
for all of its kills, it was incapable of killing the main Emacs
process from an `emacsclient` frame.  The new implementation allows
for this behavior using a prefix argument; the ability to save all
buffers silently (the normal behavior of `save-buffer-kill-emacs` or
`save-buffers-kill-terminal` with a prefix arg) is accessible via a
double prefix argument.

Additionally:

- The docstring of this function was unclear about what exactly should
  happen when called from an `emacsclient` frame.  This is because,
  somewhat surprisingly, the invisible server frame actually does
  appear in `visible-frame-list`.  The new docstring makes the
  expected behavior more clear.

- The previous implementation "baked in" a confirmation prompt.  It
  seems better to leave that decision to the existing variable
  `confirm-kill-emacs`, and then bypass that variable in the `C-u C-u`
  case.`