New command `aph/outline-forward-end-of-level`
This command (bound to `C-M-e` in `outline-mode`) moves forward to the
last sibling of the current heading.

There is also a non-interactive function
`aph/outline-get-final-sibling`, which differs only in its approach to
the situation where the current heading has no later siblings (the
interactive command signals an error to let the user know what's
wrong, while the non-interactive function just returns nil).