New macro `aph/keys-define-conditionally`
This macro (which I've adapted from code on Artur Malabarba's blog) is
an analogue for `define-key` that also accepts arbitrary elisp code.
The binding established is only active when that code evaluates
non-nil.

Unlike Malabarba's version, I've separated out the portion of the
macro that constructs the binding form as a function
`aph/keys--construct-conditional-binding`; this should make it easier
to add `bind-key` support (my next project).