Rename `aph-keys-mode` to `umbra-mode`
The module `aph-keys` was getting too disorganized for my liking.  In
particular, it shouldn't house both `aph-keys-mode` and the code
involving conditional keybindings, which is fundamentally unrelated.

Furthermore, `aph-keys-mode` is a horrible name, and the mixed
prefixes `aph-` and `aph/` were starting to get confusing.  Hence I
have renamed `aph-keys-mode` as `umbra-mode`.  Most of the functions
involved have gotten new names, and the :augment and :override
keywords for `bind-keys` have been renamed :umbra and :penumbra,
respectively.