Clean up `umbra` dependencies
Of the four dependencies previously `require`d by `umbra`
(`aph-symbol`, `aph-bind-key`, `cl-lib`, and `aph-plist`), only two
(`aph-symbol` and `aph-plist`) are actually necessary.

The `cl-lib` dependency was for a `cl-defun` that has already been
refactored away, and is no longer necessary.

The `aph-bind-key` dependency is only for implementing `bind-key`
support (and picking up the menu item support that was necessary for
code remaining in the `aph-keys` module).  Since advice can safely be
applied to a function before that function is applied, removing this
means that `umbra` can be used without `bind-key`, while still adding
support for it when it is present.

Moreover, the `aph-plist` dependency is now only necessary when
`bind-key` is present.