Revert `umbra-mode`'s handling of `after-change-major-mode-hook`
A recent commit caused `after-change-major-mode-hook` to only hold
`umbra--update-major-mode` when `umbra-mode` is active.  This had
unintended consequences, in that it makes it more difficult to
`let`-bind `umbra-mode` in tests.  A more thorough rewrite is probably
called for, but for the moment, let's just go back to the way we were
doing things.