Rework implementation of `mode-family` module
The original implementation of `mode-family` stored its data as symbol
properties.  This worked well overall, but did not allow for the
possibility of a name collision between a mode family and an unrelated
symbol.  In the rare event that some other package uninterned that
symbol, it would delete the mode family, too.

Hence this refactoring, which replaces symbol properties with a
list-based implementation.  While we're changing things, we also made
the following changes:

- We removed the docstring functionality for mode families, as this
  had yet to be fully implemented.  We can add it again later if
  desired.

- We removed the dependency on `aph-symbol`, essentially by inlining
  `aph/symbol-concat`.