Clean up `mode-family`'s `use-package` declaration
With the recent refactoring of the `mode-family` module, it is no
longer necessary to explicitly create each mode family, so we can
remove the `mode-family-create` forms for the lisp and clojure
families.

Furthermore, adding an autoload statement for `mode-family-add` lets
us `:defer` the `mode-family` module.  In reality, this doesn't
actually get us anything, because the `lisp-mode` module is not
deferred and uses mode families, but it looks tidier.